+++ crates/alpaca-trade/src/account_configurations/request.rs
@@ -5,8 +5,6 @@ use crate::Error;
 #[derive(Clone, Debug, Default, PartialEq, Eq, Serialize)]
 pub struct UpdateRequest {
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub dtbp_check: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub trade_confirm_email: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub suspend_trade: Option<bool>,
@@ -18,8 +16,6 @@ pub struct UpdateRequest {
     pub max_margin_multiplier: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub max_options_trading_level: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub pdt_check: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub ptp_no_exception_entry: Option<bool>,
     #[serde(default, skip_serializing_if = "Option::is_none")]